--- optuna/artifact/__init__.py
@@ -1,4 +1,0 @@
-from optuna.artifact._filesystem import FileSystemArtifactStore
-
-
-__all__ = ["FileSystemArtifactStore"]
