+++ optuna/artifacts/__init__.py
@@ -1,0 +1,4 @@
+from optuna.artifacts._filesystem import FileSystemArtifactStore
+
+
+__all__ = ["FileSystemArtifactStore"]
