+++ optuna/artifacts/exceptions.py
